feat: refactored the code
